Single loop, move reusable in separate module
